MPOPF API Reference

Contents

Contents

Index

Module Documentation

Types

Functions

Index

MPOPF.MPOPF

MPOPF.ACMPOPFModelFactory

MPOPF.AbstractMPOPFModel

MPOPF.AbstractMPOPFModelFactory

MPOPF.DCMPOPFModelFactory

MPOPF.LinMPOPFModelFactory

MPOPF.MPOPFModel

MPOPF.MPOPFModelUncertainty

MPOPF.NewACMPOPFModelFactory

MPOPF.calculate_model_averages

MPOPF.compute_and_save_feasibility

MPOPF.create_model

MPOPF.create_model

MPOPF.create_model_check_feasibility

MPOPF.find_bound_violations

MPOPF.find_infeasible_constraints

MPOPF.generate_load_scenarios

MPOPF.get_ref

MPOPF.load_and_compile_models

MPOPF.load_and_compile_results

MPOPF.load_and_graph_results

MPOPF.optimize_model

MPOPF.optimize_model_with_plot

MPOPF.perform_feasibility

Module Documentation

MPOPF.MPOPF — Module

MPOPF

A module for Multi-Period Optimal Power Flow (MPOPF) modeling and optimization.

This module provides tools to create, optimize, and analyze MPOPF models using various formulations including AC, DC, and linearized versions.

Main Features

Create MPOPF models using different factory types (AC, DC, Linear, etc.)

Optimize MPOPF models

Handle uncertainty in MPOPF models

Visualize optimization results

Types

MPOPF.ACMPOPFModelFactory — Type

ACMPOPFModelFactory <: AbstractMPOPFModelFactory

Factory for creating AC MPOPF models.

Fields

file_path::String: Path to the input data file.

optimizer::Type: The optimizer to be used (e.g., Ipopt.Optimizer).

MPOPF.AbstractMPOPFModel — Type

AbstractMPOPFModel

An abstract type serving as a base for all MPOPF model types.

MPOPF.AbstractMPOPFModelFactory — Type

AbstractMPOPFModelFactory

An abstract type serving as a base for all MPOPF model factories.

MPOPF.DCMPOPFModelFactory — Type

DCMPOPFModelFactory <: AbstractMPOPFModelFactory

Factory for creating DC MPOPF models.

Fields

file_path::String: Path to the input data file.

optimizer::Type: The optimizer to be used.

MPOPF.LinMPOPFModelFactory — Type

LinMPOPFModelFactory <: AbstractMPOPFModelFactory

Factory for creating linearized MPOPF models. If this factory is used then model_type needs to be specified in the create_model() function.

Fields

file_path::String: Path to the input data file.

optimizer::Type: The optimizer to be used.

MPOPF.MPOPFModel — Type

MPOPFModel <: AbstractMPOPFModel

Represents a Multi-Period Optimal Power Flow model.

Fields

model::JuMP.Model: The underlying JuMP model.

data::Dict: Dictionary containing the power system data.

time_periods::Int64: Number of time periods in the model.

factors::Vector{Float64}: Scaling factors for each time period.

ramping_cost::Int64: Cost associated with generator ramping.

MPOPF.MPOPFModelUncertainty — Type

MPOPFModelUncertainty <: AbstractMPOPFModel

Represents a Multi-Period Optimal Power Flow model with uncertainty considerations.

Fields

model::JuMP.Model: The underlying JuMP model.

data::Dict: Dictionary containing the power system data.

scenarios::Dict: Dictionary of scenarios for uncertainty analysis.

time_periods::Int64: Number of time periods in the model.

factors::Vector{Float64}: Scaling factors for each time period.

ramping_cost::Int64: Cost associated with generator ramping.

MPOPF.NewACMPOPFModelFactory — Type

NewACMPOPFModelFactory <: AbstractMPOPFModelFactory

Factory for creating new AC MPOPF models. This is only used when computing for feasibility

Fields

file_path::String: Path to the input data file.

optimizer::Type: The optimizer to be used.

Functions

MPOPF.calculate_model_averages — Method

calculate_model_averages(model_results::Dict{String, Dict{String, Dict{String, Float64}}})

Calculate the average values for cost, verror, oerror and times for all models in the dictionary

Arguments

model_results::Dict{String, Dict{String, Dict{String, Float64}}}: The dictionary of models to be analysed.

This model_results can be obtained by running the function load_and_compile_results.

Returns

Dict{String, Dict{String, Float64}}: A dictionary that maps the model to a metric to a value

MPOPF.compute_and_save_feasibility — Function

compute_and_save_feasibility(factory::Union{ACMPOPFModelFactory, DCMPOPFModelFactory, LinMPOPFModelFactory}, path::String, model_type=false)

Compute the feasibility of a model and save the results to a specified path.

Arguments

factory::Union{ACMPOPFModelFactory, DCMPOPFModelFactory, LinMPOPFModelFactory}: The factory to create the model.

path::String: The path where the results will be saved.

model_type: The type of the model (optional, only needed for LinMPOPFModelFactory).

Returns

Dict: A dictionary containing the results of the computation.

MPOPF.create_model — Function

create_model(factory::AbstractMPOPFModelFactory, scenarios::Dict, time_periods::Int64=1, factors::Vector{Float64}=[1.0], ramping_cost::Int64=0)::MPOPFModelUncertainty

Create a Multi-Period Optimal Power Flow (MPOPF) model with uncertainty considerations.

Arguments

factory: The factory used to create the specific type of MPOPF model.

scenarios: Dictionary of scenarios for uncertainty analysis.

time_periods: Number of time periods to consider in the model. Default is 1.

factors: Scaling factors for each time period. Default is [1.0].

ramping_cost: Cost associated with ramping generation up or down. Default is 0.

Returns

An MPOPFModelUncertainty object representing the created MPOPF model with uncertainty.

MPOPF.create_model — Method

create_model(factory::AbstractMPOPFModelFactory; time_periods::Int64=1, factors::Vector{Float64}=[1.0], ramping_cost::Int64=0, model_type=undef)::MPOPFModel

Create a Multi-Period Optimal Power Flow (MPOPF) model. The factory passed in defines the type of model that will be returned

If LinMPOPFModelFactory is passed in as the factory then model_type needs to be specified.

Arguments

factory: The factory used to create the specific type of MPOPF model.

time_periods: Number of time periods to consider in the model. Default is 1.

factors: Scaling factors for each time period. Default is [1.0].

ramping_cost: Cost associated with ramping generation up or down. Default is 0.

model_type: Optional parameter to specify a particular model type. Default is undef.

Returns

An MPOPFModel object representing the created MPOPF model.

MPOPF.create_model_check_feasibility — Function

create_model_check_feasibility(factory::NewACMPOPFModelFactory, new_pg=false, new_qg=false, v=false, theta=false, time_periods::Int64=1, factors::Vector{Float64}=[1.0], ramping_cost::Int64=0)::MPOPFModel

Create a secondary model to assess the feasibility of a solution when comparing to the AC model.

Arguments

factory: The factory used to create the specific type of MPOPF model.

new_pg: Fixed values for active power generation, or false to skip fixing.

new_qg: Fixed values for reactive power generation, or false to skip fixing.

v: Fixed values for bus voltages, or false to skip fixing.

theta: Fixed values for bus angles, or false to skip fixing.

time_periods: Number of time periods to consider in the model. Default is 1.

factors: Scaling factors for each time period. Default is [1.0].

ramping_cost: Cost associated with ramping generation up or down. Default is 0.

Returns

An MPOPFModel object representing the created MPOPF model.

MPOPF.find_bound_violations — Method

find_bound_violations(model::Model)

Check lower and upper bounds of all variables in a model, returns a dictionary of violations with additional mismatch data

Arguments

model::Model: The model to be checked for vounds violations

Returns

Dict{VariableRef, Tuple{Float64, Float64, Float64, Float64}}: A dictionary where each key

is a variable ref and the value is a tuple of data.

MPOPF.find_infeasible_constraints — Method

find_infeasible_constraints(model::Model)

check the feasibility of the constraints in a model, returns a dictionary of infeasible constraints

Arguments

model::Model: The model to be checked for constraint violations

Returns

[]: An array of the infeasible constraints.

MPOPF.generate_load_scenarios — Method

generate_load_scenarios(num_scenarios::Int, num_buses::Int)

Generate load scenarios for a given number of scenarios and buses.

Arguments

num_scenarios::Int: The number of scenarios to generate.

num_buses::Int: The number of buses in the case file.

Returns

Dict{Int, Dict{Int, Float64}}: A dictionary where each key

is a scenario number and the value is another dictionary mapping bus numbers to load factors.

MPOPF.get_ref — Method

get_ref(data::Dict{String, Any})

Build and return a reference object from the given power system data dictionary.

Arguments

data: Dictionary containing the power system data for a given case file.

Returns

A reference object containing processed power system data.

MPOPF.load_and_compile_models — Method

load_and_compile_models(results_directory::String)

Load models (Contrary to load_and_compile_results) from a directory and compile them into a dictionary for further analysis. Example, getting the pg value from case14 AC: model_results["case14"]["AC"].model[:pg].

Arguments

results_directory::String: The directory where the models are stored.

Returns

Dict{String, Dict{String, Dict{String, Any}}}: A dictionary containing the data of the models.

MPOPF.load_and_compile_results — Function

load_and_compile_results(results_directory::String, save_to_file::Bool=false)

Load results from a directory and compile them into a dictionary for further analysis.

Arguments

results_directory::String: The directory where the results are stored. To get results run the function compute_and_save_feasibility.

save_to_file::Bool: Whether to save the generated results to a file (optional) (Not yet implemented).

Returns

Dict{String, Dict{String, Dict{String, Float64}}}: A dictionary containing the results of the computation.

MPOPF.load_and_graph_results — Function

load_and_graph_results(results_directory::String, save_to_file::Bool=false)

Load results from a directory and generate graphs for feasibility, voltage magnitude error, voltage angle error, and computation times (times not complete).

Arguments

results_directory::String: The directory where the results are stored. To get results run the function compute_and_save_feasibility.

save_to_file::Bool: Whether to save the generated graphs to a file (optional).

Returns

feasibility_graph, v_error_graph, o_error_graph, and time_graph.

MPOPF.optimize_model — Method

optimize_model(model::AbstractMPOPFModel)

Optimize the given MPOPF model and print the optimal cost.

Note

If the model being optimized is of type MPOPFModelUncertainty then extra computation will be executed to determine if there are any unbalaced power at each bus for every scenario when consitering muplus and muminus.

That is if sum_of_mu_plus and mu_minus are both > 0.01 at any given bus then an error message is printed.

Arguments

model: The MPOPF model to optimize.

MPOPF.optimize_model_with_plot — Method

optimize_model_with_plot(model::AbstractMPOPFModel)

Optimize the given MPOPF model, print the optimal cost, and generate a PlotlyJS graph/plot of the optimization process.

Note

if time_periods = 1 then the graph will represent the iterations taken by the solver otherwize the graph will represent each time period.

Arguments

model: The MPOPF model to optimize and plot.

MPOPF.perform_feasibility — Function

perform_feasibility(models::Array, finish_save::Bool=false)

Fields

models::Array : a vector of 5, where each element is a model to be performed

- 1 = AC, 2 = DC, 3 = Lin1, 4 = Lin2, 5 = Lin3
- toggle 1 = on, 0 = off

finish_save::Bool=false : a boolean to determine if the graphs should be saved

Description

This function allows the user to execute feasibility tests on the different models provided

- AC
- DC
- Lin1
- Lin2
- Lin3

for each case in the cases folder the function will loop over them and perform the feasibility tests

for each model that has been toggled to be on